refactor: split ui/main_window.py (3909 lines) into 10 mixin modules
Decomposed monolithic MainWindow class using Python mixins:
- menu_mixin.py: menu bar construction
- shortcuts_mixin.py: keyboard shortcuts, view mode hotkeys
- clip_mixin.py: clip selection, directory scanning
- import_mixin.py: folder/video/sequence import flows
- inference_mixin.py: inference control, job submission
- worker_mixin.py: GPU worker signal handlers
- annotation_mixin.py: paint annotations, SAM2 tracking
- export_mixin.py: extraction, video export, in/out markers
- session_mixin.py: session save/load persistence
- settings_mixin.py: preferences, about, update check

main_window.py reduced from 3909 to 778 lines.
All 427 tests pass.
